+++ lib/vagrant/action/vm/provision.rb
@@ -1,14 +1,18 @@
+require "log4r"
+
 module Vagrant
   module Action
     module VM
       class Provision
         def initialize(app, env)
+          @logger = Log4r::Logger.new("vagrant::action::vm::provision")
           @app = app
-          @env = env
-          @env["provision.enabled"] = true if !@env.has_key?("provision.enabled")
+          @env["provision.enabled"] = true if !env.has_key?("provision.enabled")
         end
 
         def call(env)
+          @env = env
+
           provisioners = nil
 
           # We set this here so that even if this value is changed in the future,
@@ -34,9 +38,23 @@ module Vagrant
         end
 
         def enabled_provisioners
-          @env[:vm].config.vm.provisioners.map do |provisioner|
-            provisioner.provisioner.new(@env, provisioner.config)
+          enabled = []
+          @env[:vm].config.vm.provisioners.each do |provisioner|
+            if @env["provision.types"]
+              # If we've specified types of provisioners to enable, then we
+              # only use those provisioners, and skip any that we haven't
+              # specified.
+              if !@env["provision.types"].include?(provisioner.shortcut.to_s)
+                @logger.debug("Skipping provisioner: #{provisioner.shortcut}")
+                next
+              end
+            end
+
+            enabled << provisioner.provisioner.new(@env, provisioner.config)
           end
+
+          # Return the enable provisioners
+          enabled
         end
       end
     end