+++ lib/vagrant/command/up.rb
@@ -4,7 +4,10 @@ module Vagrant
   module Command
     class Up < Base
       def execute
-        options = { :provision => true }
+        options = {
+          :provision => true,
+          :provisioners => nil
+        }
 
         opts = OptionParser.new do |opts|
           opts.banner = "Usage: vagrant up [vm-name] [--[no-]provision] [-h]"
@@ -14,11 +17,22 @@ module Vagrant
           opts.on("--[no-]provision", "Enable or disable provisioning") do |p|
             options[:provision] = p
           end
+
+          opts.on("--provision-with x,y,z", Array,
+                  "Enable only certain provisioners, by type.") do |list|
+            options[:provisioners] = list
+          end
         end
 
         # Parse the options
         argv = parse_options(opts)
         return if !argv
+
+        # Parameters to send to actions
+        action_params = {
+          "provision.enabled" => options[:provision],
+          "provision.types"   => options[:provisioners]
+        }
 
         # Go over each VM and bring it up
         @logger.debug("'Up' each target VM...")
@@ -26,10 +40,10 @@ module Vagrant
           if vm.created?
             @logger.info("Booting: #{vm.name}")
             vm.ui.info I18n.t("vagrant.commands.up.vm_created")
-            vm.start("provision.enabled" => options[:provision])
+            vm.start(action_params)
           else
             @logger.info("Creating: #{vm.name}")
-            vm.up("provision.enabled" => options[:provision])
+            vm.up(action_params)
           end
         end
       end