+++ lib/vagrant/provisioners/shell.rb
@@ -82,13 +82,15 @@ module Vagrant
           env[:vm].channel.upload(path.to_s, config.upload_path)
 
           # Execute it with sudo
-          env[:vm].channel.sudo(command) do |ch, type, data|
-            # Output the data with the proper color based on the stream.
-            color = type == :stdout ? :green : :red
+          env[:vm].channel.sudo(command) do |type, data|
+            if [:stderr, :stdout].include?(type)
+              # Output the data with the proper color based on the stream.
+              color = type == :stdout ? :green : :red
 
-            # Note: Be sure to chomp the data to avoid the newlines that the
-            # Chef outputs.
-            env[:ui].info(data.chomp, :color => color, :prefix => false)
+              # Note: Be sure to chomp the data to avoid the newlines that the
+              # Chef outputs.
+              env[:ui].info(data.chomp, :color => color, :prefix => false)
+            end
           end
         end
       end