+++ test/acceptance/skeletons/provisioner_multi/cookbooks/basic/recipes/default.rb
@@ -1,0 +1,5 @@
+# Just create a file
+file "/vagrant/chef_solo" do
+  mode    0644
+  content "success"
+end